correct data column format and phone number data types

+++ db/migrations/schemas/dish-cost.js
@@ -1,4 +1,4 @@
 module.exports = (table) => {
   table.increments()
-  table.integer('cost')
+  table.decimal('cost')
 }
