+++ db/migrations/schemas/ingredient-dish.js
@@ -1,7 +1,7 @@
 module.exports = (table) => {
   table.increments()
   table.enu('unit', ['serving', 'ounce', 'gram', 'millimeter', 'litres', 'pound', 'teaspoon', 'tablespoon', 'cup']).notNullable()
-  table.integer('ingredient_portion').notNullable()
+  table.decimal('ingredient_portion').notNullable()
   table.integer('ingredient_id').unsigned().notNullable()
   table.foreign('ingredient_id').references('ingredient.id')
   table.integer('dish_id').unsigned().notNullable()
