+++ db/migrations/schemas/menu-dish.js
@@ -1,6 +1,6 @@
 module.exports = (table) => {
   table.increments()
-  table.integer('dish_portion').notNullable()
+  table.decimal('dish_portion').notNullable()
   table.string('meal')
   table.integer('menu_id').unsigned()
   table.foreign('menu_id').references('menu.id')
