+++ db/migrations/schemas/user-info.js
@@ -1,6 +1,6 @@
 module.exports = (table) => {
   table.increments()
   table.string('email')
-  table.integer('phone_number')
+  table.bigint('phone_number')
   table.string('sex')
 }
